+++ CHANGELOG.md
@@ -5,6 +5,7 @@
  * Don't throw errors on empty responses (204s for example are valid, empty body responses)
  * Request adapter should not send empty JSON body
  * Add support for submitting arrays as the body of an action
+ * Bug fix - don't check empty body for starting { in request adapter
 
 ### 0.0.2
  * Add `onRejected` handler for managing connectivity issues